Move code to create charts from Behaviour to Notification class.
